+++ utills/utills.js
@@ -1,0 +1,9 @@
+import jwt from "jsonwebtoken";
+
+export function generateAccessToken(payload) {
+  return jwt.sign(payload, process.env.JWT_SECRET, { expiresIn: "10m" });
+}
+
+export function generateRefreshToken(payload) {
+  return jwt.sign(payload, process.env.JWT_REFRESH_SECRET, { expiresIn: "7d" });
+}
